+++ pkg/repository/repository.go
@@ -14,8 +14,8 @@ func NewRepository(db *sql.DB) *Repository {
 }
 
 type Actor interface {
-	Add(actor *model.Actor) (int, error)
-	Update(actorId int, actor *model.Actor) (*model.Actor, error)
-	Delete(actorId int) (*model.Actor, error)
+	Add(actor *model.Actor) (int64, error)
+	Update(actorId int64, actor *model.Actor) (*model.Actor, error)
+	Delete(actorId int64) (*model.Actor, error)
 	GetAll() ([]*model.Actor, error)
 }
